fix(plugin-nlu): rename fileUtils to kebab-case
